--- mrbgems/picoruby-net/src/mbedtls_debug.c
@@ -1,8 +1,0 @@
-#include "mruby.h"
-
-void
-mbedtls_debug(void *ctx, int level, const char *file, int line, const char *str)
-{
-  ((void) level);
-  printf("%s:%04d: %s", file, line, str);
-}
